# sprinkle at-inline and at-propagate_inbounds around