Timings:

Before:
```
julia> using DataStructures

julia> using BenchmarkTools

julia> cb=CircularBuffer{Float64}(3)
0-element DataStructures.CircularBuffer{Float64}

julia> pusher(cb)=(@inbounds for i=1:10^6; push!(cb, 3.);end; nothing)
pusher (generic function with 1 method)

julia> @benchmark pusher(cb)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     6.944 ms (0.00% GC)
  median time:      6.995 ms (0.00% GC)
  mean time:        7.029 ms (0.00% GC)
  maximum time:     8.994 ms (0.00% GC)
  --------------
  samples:          713
  evals/sample:     1
```

After:
```
julia> @benchmark pusher(cb)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.400 ms (0.00% GC)
  median time:      2.414 ms (0.00% GC)
  mean time:        2.422 ms (0.00% GC)
  maximum time:     3.129 ms (0.00% GC)
  --------------
  samples:          2069
  evals/sample:     1
```

Similar results for larger buffers.