@oxinabox 2018-06-21
Those performance improvements are nice
Good stuff.
And to me this looks completely correct.
(I'm not super familiar with the inbound propergation macros but this match's the manual).

I'm a bit unhappy about the 0.7 tests failing, but I think that us unrelated.
I'll double check tomorrow and get back to you.

If this doesn't happen in next few days feel free to pung me. I might have forgotten.