@oxinabox 2018-06-22
For my own future reference here is a link to a build of master I just triggered.
https://travis-ci.org/JuliaCollections/DataStructures.jl/builds/395401162
If it fails on 0.7 then that means this PR didn't break 0.7 and so i think am all clear to merge it.