@oxinabox 2018-06-22
Yes, master is failing, and in the same way.
And this code looks right, so this is all good.

Thank you for this,